Update CryptTreeNode to use CryptTreeNodeModel
